feat : 가게 도메인 dto 경로 변경
