+++ src/main/java/com/nextdaydelivery/store/presentation/dto/request/StoreUpdateRequest.java
@@ -1,0 +1,14 @@
+package com.nextdaydelivery.store.presentation.dto.request;
+
+import java.util.List;
+import java.util.UUID;
+
+public record StoreUpdateRequest(
+    String name,
+    String sido,
+    String sigungu,
+    String dong,
+    String detailAddress,
+    List<UUID> categoryIds
+) {
+}
